Repo added

+++ src/main/java/com/noteManager/model/Note.java
@@ -20,7 +20,7 @@ import static jakarta.persistence.GenerationType.IDENTITY;
 public class Note implements BaseEntity{
     @Id
     @GeneratedValue(strategy = IDENTITY)
-    private double id;
+    private Long id;
 
     @NotBlank
     private String title;
